change data type of position in troop drop struct to int from uint8

+++ backend/internal/models/battle_model.go
@@ -8,8 +8,8 @@ import (
 
 type TroopDropBody struct {
 	Name string `json:"name"`
-	X uint8 `json:"x"`
-	Y uint8 `json:"y"`
+	X int `json:"x"`
+	Y int `json:"y"`
 }
 type BattleManager struct {
     Mu      *sync.Mutex
